[#1474] Use AtomicU64 as basis for low level concurrency constructs

+++ iceoryx2-bb/concurrency/src/internal/strategy/barrier.rs
@@ -15,7 +15,7 @@
 pub struct Barrier(internal::Barrier);
 
 impl Barrier {
-    pub fn new(n: u16) -> Self {
+    pub fn new(n: u32) -> Self {
         Self(internal::Barrier::new(n))
     }
 }
