+++ iceoryx2-pal/concurrency-sync/src/strategy/condition_variable.rs
@@ -12,14 +12,14 @@
 
 pub use crate::strategy::mutex::Mutex;
 
-use crate::atomic::AtomicU32;
+use crate::atomic::AtomicU64;
 use crate::atomic::Ordering;
 use crate::strategy::semaphore::Semaphore;
 use crate::{WaitAction, WaitResult};
 
 #[derive(Debug)]
 pub struct ConditionVariable {
-    number_of_waiters: AtomicU32,
+    number_of_waiters: AtomicU64,
     semaphore: Semaphore,
 }
 
@@ -27,7 +27,7 @@ impl Default for ConditionVariable {
     fn default() -> Self {
         Self {
             semaphore: Semaphore::new(0),
-            number_of_waiters: AtomicU32::new(0),
+            number_of_waiters: AtomicU64::new(0),
         }
     }
 }
@@ -37,22 +37,22 @@ impl ConditionVariable {
         Self::default()
     }
 
-    pub fn notify_one<WakeOne: Fn(&AtomicU32)>(&self, wake_one: WakeOne) {
+    pub fn notify_one<WakeOne: Fn(&AtomicU64)>(&self, wake_one: WakeOne) {
         self.semaphore.post(
             wake_one,
             1.min(self.number_of_waiters.load(Ordering::Relaxed)),
         );
     }
 
-    pub fn notify_all<WakeAll: Fn(&AtomicU32)>(&self, wake_all: WakeAll) {
+    pub fn notify_all<WakeAll: Fn(&AtomicU64)>(&self, wake_all: WakeAll) {
         self.semaphore
             .post(wake_all, self.number_of_waiters.load(Ordering::Relaxed));
     }
 
     pub fn wait<
-        WakeOne: Fn(&AtomicU32),
-        Wait: Fn(&AtomicU32, &u32) -> WaitAction,
-        MtxWait: Fn(&AtomicU32, &u32) -> WaitAction,
+        WakeOne: Fn(&AtomicU64),
+        Wait: Fn(&AtomicU64, &u64) -> WaitAction,
+        MtxWait: Fn(&AtomicU64, &u64) -> WaitAction,
     >(
         &self,
         mtx: &Mutex,
